+++ libs/coin-modules/coin-stellar/package.json
@@ -134,7 +134,7 @@
     "@faker-js/faker": "^8.4.1",
     "@types/invariant": "^2.2.2",
     "@types/jest": "^29.5.10",
-    "@types/node": "^16.11.7",
+    "@types/node": "^20.8.10",
     "jest": "^29.7.0",
     "ts-jest": "^29.1.1"
   }
